% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{campaign_long}
\alias{campaign_long}
\title{Long-format Data on Negative Campaign Advertising in US Senate and Gubernatorial Elections}
\format{An object of class \code{data.frame} with 565 rows and 24 columns.}
\usage{
campaign_long
}
\description{
A dataset containing 21 columns (including treatment, outcome, id, time, and baseline
and time-varying confounders) and 565 unit-week records from Blackwell (2013).
}
\references{
Blackwell, Matthew. 2013. A Framework for Dynamic Causal Inference in
  Political Science. American Journal of Political Science 57(2): 504-619.
}
\keyword{datasets}
