% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{education}
\alias{education}
\title{Data on College Education and Depression among US adults}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2924 rows and 18 columns.}
\usage{
education
}
\description{
A dataset containing 18 columns (including treatment, mediator, outcome, and
baseline and post-treatment confounders) and 2,924 individuals.
}
\keyword{datasets}
