% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbwMed.R
\name{rbwMed}
\alias{rbwMed}
\title{Residual Balancing Weights for Causal Mediation Analysis}
\usage{
rbwMed(treatment, mediator, zmodels, baseline_x, base_weights, data,
  max_iter = 500, print_level = 1, tol = 0.001)
}
\arguments{
\item{treatment}{Expression for the treatment variable.}

\item{mediator}{Expression for the mediator variable.}

\item{zmodels}{A list of fitted \code{lm} or \code{glm} objects for
post-treatment confounders of the mediator-outcome relationship.}

\item{baseline_x}{Expression for a set of baseline confounders. .}

\item{base_weights}{Expression for base weights (optional).}

\item{data}{A data frame containing the variables in the model.}

\item{max_iter}{Maximum number of iterations for Newton's method.}

\item{print_level}{The level of printing: 1 (normal printing), 2 (detailed),
and 3 (very detailed).}

\item{tol}{Tolerance parameter used to determine convergence.}
}
\value{
A list containing the results.
 \item{weights}{A vector of residual balancing weights.}
 \item{constraints}{A matrix of (linearly independent) residual balancing constraints}
 \item{eb_out}{Results from calling \code{\link{eb2}} function}
 \item{call}{The matched call.}
}
\description{
\code{rbwMed} is a function that produces residual balancing weights for
causal mediation analysis. The weights can be used to fit marginal
structural models for the joint effects of the treatment and a mediator.
}
\examples{
# models for post-treatment confounders
m1 <- lm(cesd92 ~ male + black + test_score + educ_exp +  father +
  hispanic + urban + educ_mom + num_sibs + college, data = education)
m2 <- lm(prmarr98 ~ male + black + test_score + educ_exp +  father +
  hispanic + urban + educ_mom + num_sibs + college, data = education)
m3 <- lm(transitions98 ~ male + black + test_score + educ_exp + father +
  hispanic +urban + educ_mom + num_sibs + college, data = education)

# residual balancing weights
rbwMed_fit <- rbwMed(treatment = college, mediator = ses, baseline_x = male:num_sibs,
  zmodels = list(m1, m2, m3), base_weights = weights, data = education)
summary(rbwMed_fit$weights)
}
