% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbwMed.R
\name{rbwMed}
\alias{rbwMed}
\title{Residual Balancing Weights for Causal Mediation Analysis}
\usage{
rbwMed(
  treatment,
  mediator,
  zmodels,
  data,
  baseline_x,
  base_weights,
  max_iter = 200,
  print_level = 1,
  tol = 1e-04
)
}
\arguments{
\item{treatment}{Symbol for the treatment variable.}

\item{mediator}{Symbol for the mediator variable.}

\item{zmodels}{A list of fitted \code{lm} or \code{glm} objects for
post-treatment confounders of the mediator-outcome relationship.}

\item{data}{A data frame containing all variables in the model.}

\item{baseline_x}{(Optional) Expression for a set of baseline confounders.}

\item{base_weights}{(Optional) Symbol for base weights.}

\item{max_iter}{Maximum number of iterations for Newton's method.}

\item{print_level}{The level of printing: 1 (normal printing), 2 (detailed),
and 3 (very detailed).}

\item{tol}{Tolerance parameter used to determine convergence.}
}
\value{
A list containing the results.
 \item{weights}{A vector of residual balancing weights.}
 \item{constraints}{A matrix of (linearly independent) residual balancing constraints}
 \item{eb_out}{Results from calling \code{\link{eb2}} function}
 \item{call}{The matched call.}
}
\description{
\code{rbwMed} is a function that produces residual balancing weights for
causal mediation analysis. The weights can be used to fit a marginal
structural model for the joint effects of treatment and a mediator
on the outcome of interest.
}
\examples{
# models for post-treatment confounders
m1 <- lm(cesd92 ~ male + black + test_score + educ_exp +  father +
  hispanic + urban + educ_mom + num_sibs + college, weights = weights, data = education)
m2 <- lm(prmarr98 ~ male + black + test_score + educ_exp +  father +
  hispanic + urban + educ_mom + num_sibs + college, weights = weights, data = education)
m3 <- lm(transitions98 ~ male + black + test_score + educ_exp + father +
  hispanic +urban + educ_mom + num_sibs + college, weights = weights, data = education)

# residual balancing weights
rbwMed_fit <- rbwMed(treatment = college, mediator = ses, baseline_x = male:num_sibs,
  zmodels = list(m1, m2, m3), base_weights = weights, data = education)
summary(rbwMed_fit$weights)

# attach residual balancing weights to data
education$rbw <- rbwMed_fit$weights

# fit marginal structural model
if(require(survey)){
  rbw_design <- svydesign(ids = ~ 1, weights = ~ rbw, data = education)
  msm_rbw <- svyglm(cesd40 ~ college * ses, design = rbw_design)
  summary(msm_rbw)
}
}
